Add inline documentation
