--- dynamiclinks/DynamicLinksSwiftUIExample/AppDelegate.swift
@@ -1,31 +1,0 @@
-// Copyright 2022 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import UIKit
-import FirebaseCore
-import FirebaseDynamicLinks
-
-class AppDelegate: NSObject, UIApplicationDelegate, ObservableObject {
-  @Published var linkAlertMessage: String = ""
-  @Published var showAlertMessage: Bool = false
-
-  func application(_ application: UIApplication,
-                   didFinishLaunchingWithOptions launchOptions: [UIApplication
-                     .LaunchOptionsKey: Any]? = nil) -> Bool {
-    FirebaseApp.configure()
-    DynamicLinks.performDiagnostics()
-
-    return true
-  }
-}
